--- src/main/kotlin/community/whatever/onembackendkotlin/UrlShortenService.kt
@@ -1,7 +1,0 @@
-package community.whatever.onembackendkotlin
-
-interface UrlShortenService {
-
-    fun getOriginUrl(request: ShortenUrlSearchRequest): OriginUrlResponse
-    fun saveShortenUrl(request: ShortenUrlCreateRequest): ShortenedUrlResponse
-}
