+++ src/main/kotlin/community/whatever/onembackendkotlin/common/dto/UrlShortenResponse.kt
@@ -1,4 +1,4 @@
-package community.whatever.onembackendkotlin
+package community.whatever.onembackendkotlin.common.dto
 
 data class ShortenedUrlResponse(val shortenedUrl: String)
