+++ src/main/kotlin/community/whatever/onembackendkotlin/domain/ShortenedUrl.kt
@@ -1,3 +1,3 @@
-package community.whatever.onembackendkotlin
+package community.whatever.onembackendkotlin.domain
 
 data class ShortenedUrl(val originUrl: String, val id: String? = null)
