+++ src/main/kotlin/community/whatever/onembackendkotlin/domain/ShortenedUrlRepository.kt
@@ -1,4 +1,4 @@
-package community.whatever.onembackendkotlin
+package community.whatever.onembackendkotlin.domain
 
 interface ShortenedUrlRepository {
     /**
