+++ src/test/kotlin/community/whatever/onembackendkotlin/presentation/UrlShortenControllerTest.kt
@@ -1,6 +1,12 @@
-package community.whatever.onembackendkotlin
+package community.whatever.onembackendkotlin.presentation
 
 import com.fasterxml.jackson.databind.ObjectMapper
+import community.whatever.onembackendkotlin.application.UrlShortenService
+import community.whatever.onembackendkotlin.common.dto.OriginUrlResponse
+import community.whatever.onembackendkotlin.common.dto.ShortenUrlCreateRequest
+import community.whatever.onembackendkotlin.common.dto.ShortenUrlSearchRequest
+import community.whatever.onembackendkotlin.common.dto.ShortenedUrlResponse
+import community.whatever.onembackendkotlin.common.exception.UrlNotFoundException
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.mockito.BDDMockito.given
